Decimal degree grid from UTM projected spatial objects

Outputs a list of grid lines that can be used to plot a grid on maps. Requires a spatial object projected using UTM coordinates

deg_grid(dat, round.lat = FALSE, n.lat.grid = 3, round.lon = FALSE,
  n.lon.grid = 3, n.points = 30)

Arguments

dat
basemapLimits object

round.lat
specifying the level of rounding to be used to plot latitude grid lines. Overrides n.lat.grid

n.lat.grid
number of latitude grid lines. Alternatively use round.lat

round.lon
specifying the level of rounding to be used to plot longitude grid lines. Overrides n.lon.grid

n.lon.grid
number of longitude grid lines. Alternatively use round.lon

n.points
number of points used in creating the grid lines. The more points, the smoother the lines.

Value

Returns a UTM coodrinate grid.

Details

This is an internal function, which is automatically run by the basemap function. Common users do not need to worry about these details.

See also

basemap