Transform spatial coordinates to another projection

The function transforms spatial coordinates from original projection (decimal degrees assumed) to another projection.

transform_coord(x = NULL, lon = "longitude", lat = "latitude",
  new.names = c("lon.utm", "lat.utm"),
  proj.og = "+proj=longlat +datum=WGS84", proj.out = "+init=epsg:32633",
  map.type = NULL, verbose = TRUE, bind = FALSE)

Arguments

x
data frame to be transformed. Can be omitted if numeric vectors are assigned to lon and lat.

lon
either a name of the longitude column in x or a numeric vector containing longitude coordinates.

lat
either a name of the latitude column in x or a numeric vector containing latitude coordinates.

new.names
a vector of length 2 specifying the names of transformed longitude and latitude columns, respectively. NULL returns column names from x

proj.og
original proj4string projection. If NULL, the projection is taken from x. x must be a Spatial object in that case.

proj.out
the proj4string projection the coordinates should be transformed to. Defaults to the "svalbard" shape file projection.

map.type
a character string specifying the map type for which coodrinates should be transformed to. If NULL (default), proj.out is used to determine the returned projection. See type argument for possible map types. Overrides proj.out.

verbose
if TRUE (default), the function prints information about the returned data frame. Switch to FALSE to make the function silent.

bind
Should only transformed coordinates be returned (FALSE, default) or should x be returned with transformed coordinates (TRUE)?

Value

Returns a data frame with transformed spatial coordinates